requirements.txt
apache-airflow==2.6.2
pandas
apache-airflow-providers-postgres
psycopg2-binary==2.9.6
sqlalchemy==1.4.48
fastapi
uvicorn[standard]